@@ -16,39 +16,39 @@
 	},
 	"devDependencies": {
 		"@app/boring-avatars": "workspace:^",
-		"@playwright/test": "^1.51.0",
+		"@playwright/test": "^1.51.1",
 		"@softwaretechnik/dbml-renderer": "^1.0.30",
 		"@sveltejs/adapter-auto": "^3.3.1",
 		"@sveltejs/adapter-cloudflare": "^4.9.0",
-		"@sveltejs/adapter-cloudflare-workers": "^2.8.0",
+		"@sveltejs/adapter-cloudflare-workers": "^2.9.0",
 		"@sveltejs/adapter-vercel": "^5.6.3",
-		"@sveltejs/kit": "^2.19.0",
+		"@sveltejs/kit": "^2.20.3",
 		"@sveltejs/vite-plugin-svelte": "^4.0.4",
 		"@types/eslint": "^9.6.1",
 		"@types/pg": "^8.11.11",
 		"autoprefixer": "^10.4.21",
-		"drizzle-kit": "^0.30.5",
-		"eslint": "^9.22.0",
+		"drizzle-kit": "^0.30.6",
+		"eslint": "^9.23.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.46.1",
 		"globals": "^15.15.0",
 		"postcss": "^8.5.3",
 		"prettier": "^3.5.3",
 		"prettier-plugin-svelte": "^3.3.3",
-		"svelte": "^5.22.6",
+		"svelte": "^5.25.6",
 		"svelte-check": "^3.8.6",
 		"tailwindcss": "^3.4.17",
 		"typescript": "^5.8.2",
-		"typescript-eslint": "^8.26.1",
-		"vite": "^5.4.14",
-		"vitest": "^2.1.9"
+		"typescript-eslint": "^8.29.0",
+		"vite": "^5.4.17",
+		"vitest": "^3.1.1"
 	},
 	"type": "module",
 	"dependencies": {
 		"@app/db": "workspace:^",
 		"@fontsource-variable/inter": "^5.2.5",
 		"@fontsource/bebas-neue": "^5.2.5",
-		"@node-rs/argon2": "^1.8.3",
+		"@node-rs/argon2": "^2.0.2",
 		"@o7/icon": "^0.2.9",
 		"@oslojs/crypto": "^1.0.1",
 		"@oslojs/encoding": "^1.1.0",
@@ -61,12 +61,12 @@
 		"boring-avatars": "^1.11.2",
 		"clsx": "^2.1.1",
 		"dayjs": "^1.11.13",
-		"drizzle-orm": "^0.40.0",
-		"drizzle-zod": "^0.7.0",
+		"drizzle-orm": "^0.40.1",
+		"drizzle-zod": "^0.7.1",
 		"focus-trap": "^7.6.4",
 		"postgres": "^3.4.5",
 		"prettier-plugin-tailwindcss": "^0.6.11",
-		"resend": "^4.1.2",
+		"resend": "^4.2.0",
 		"tailwind-merge": "^2.6.0",
 		"zod": "^3.24.2"
 	}
