@@ -23,22 +23,22 @@
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.3.1",
-		"@sveltejs/kit": "^2.17.2",
+		"@sveltejs/kit": "^2.20.3",
 		"@sveltejs/package": "^2.3.10",
 		"@sveltejs/vite-plugin-svelte": "^4.0.4",
 		"@types/eslint": "^9.6.1",
-		"eslint": "^9.20.1",
+		"eslint": "^9.23.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.46.1",
 		"globals": "^15.15.0",
-		"prettier": "^3.5.1",
+		"prettier": "^3.5.3",
 		"prettier-plugin-svelte": "^3.3.3",
 		"publint": "^0.1.16",
-		"svelte": "^5.20.2",
+		"svelte": "^5.25.6",
 		"svelte-check": "^3.8.6",
-		"typescript": "^5.7.3",
-		"typescript-eslint": "^8.24.1",
-		"vite": "^5.4.14"
+		"typescript": "^5.8.2",
+		"typescript-eslint": "^8.29.0",
+		"vite": "^5.4.17"
 	},
 	"type": "module",
 	"dependencies": {
